+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `tt status`: displays the mode of the instance.
 
+### Added
+
+- `tt status`: added `pretty` option for pretty-formatted table output.
+
 ### Fixed
 
 - `tt clean` no longer tries to clean files multiple times.